update command model and conflict manager

+++ esir.dom11.nsoc-core/esir.dom11.nsoc.model/src/main/java/esir/dom11/nsoc/model/Command.java
@@ -77,7 +77,7 @@ public class Command {
         return _timeOut;
     }
 
-    public void setTimeOut(int timeOut) {
+    public void setTimeOut(long timeOut) {
         _timeOut = timeOut;
     }
